small fix to test paths and typo
